Add the App Engine compatible RTree and a README describing the basic functionalities.
Don't bother with the delete function for now - it's still buggy. I'll fix
it tomorrow, but we won't need it for the application anyway.

Another word of warning: Building the RTree is slow, as everything happens
within the datastore. I may be able to do bulk loading a bit smarter,
bear with me.