Now the actual list of result rectangles is drawn - plus the start and end point in red.
For now the result is just "all the rectangles" intersecting the large
rectangle formed by end and start.

Also I changed the input file to a subset of the CA streets.

This works pretty well. The only thing missing for functionality is
the actual algorithm that computes the route from the pre-result
rectangles.

That's going to be my last commit for today. I'll do some clean up
tomorrow and add in the last bits and pieces.