Turns out, once deployed we are struggling with timeouts.
It takes a moment to build an R-Tree and that moment is already too long
for the "live" version. Anyway, I put in a function that reads the R-Tree from
a pickle file and continues from there. That works for 20-30k entries. If
we want more, I need to divide and conquer a bit more.

Anyway, that gives us a decent idea of how it works already. This version is
now deployed as well.