Integrate Aaron's path algorithm.
The fact that build_graph is quadratic in size of the candidate group seems
to perform better than repeated RTree searches. So I'll leave that for now.

Nevertheless ... be prepared for this to be slow. I'll tweak it a bit more later and then deploy it to the server (right now I suspect that we'll see timeouts
for the deployed version).

But on the plus side: It seems to basically work.